+++ src/Delizious.Toolkit/Match.cs
@@ -64,6 +64,18 @@
         public static Match<T> Null<T>() where T : class
             => Match<T>.Create(NullMatch<T>.Create());
 
+        /// <summary>
+        /// Creates a <see cref="Match{T}"/> instance that matches successfully when the value to match is not a <c>null</c> reference.
+        /// </summary>
+        /// <typeparam name="T">
+        /// The type of the value to match. This must be a reference type.
+        /// </typeparam>
+        /// <returns>
+        /// A new <see cref="Match{T}"/> instance that determines whether the value to match is not a <c>null</c> reference.
+        /// </returns>
+        public static Match<T> NotNull<T>() where T : class
+            => Match<T>.Create(NotMatch<T>.Create(NullMatch<T>.Create()));
+
         private sealed class NullMatch<T> : IMatch<T>
             where T : class
         {